Add files via upload
American Bank Dashboard will help to know about the customers of Bank. It is the detailed analysis of the customer on the basis of Age, Gender, Education Balance, Monthly Balance of their account and also it gives the information about the Bank such as Balance of the Bank, No of Branches in different states. It emphasizes on the Loan details of the customer on the basis of the different parameters with the help of tooltip feature. Although being the simplest dashboard giving enough information to analyze the customers on different parameters .